Adding user registration, including validation and password encryption

+++ admin/includes/view_all_comments.php
@@ -59,7 +59,7 @@
                 echo "<td>{$comment_date}</td>";
                 echo "<td><a href='comments.php?approve={$comment_id}'>APPROVE</a></td>";
                 echo "<td><a href='comments.php?unapprove={$comment_id}'>UNAPPROVE</a></td>";
-                echo "<td><a href='comments.php?delete={$comment_id}'>DELETE</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='comments.php?delete={$comment_id}'>DELETE</a></td>";
                 echo "</tr>";
 
             }
